--- frontend/src/pages/user/User.vue
@@ -1,8 +1,0 @@
-<script setup>
-</script>
-
-<template>
-  <div>
-    LOL
-  </div>
-</template>
